gomod: extract go.py and test_go.py
Move Go, GoVersion, GoWork, ParsedGoWork, _GoWorkUseStruct,
HERMETO_GO_INSTALL_DIR, _get_gomod_version, _select_toolchain,
_list_installed_toolchains, _get_go_work_path, _list_toolchain_files
into go.py. Extract their tests into test_go.py. Update mock paths.
